pyproject.toml
[tool.poetry]
name = "rl-implement"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.12"
torch = "^2.4.1"
gymnasium = {extras = ["box2d"], version = "^0.29.1"}
numpy = "^2.1.1"
matplotlib = "^3.9.2"
plotly = "^5.24.1"
pandas = "^2.2.2"
seaborn = "^0.13.2"
mujoco = "^3.2.3"
imageio = "^2.35.1"
colorama = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
